lan: fix lan print fails on unsupported parameters
After upgrading to ipmitool 1.8.19, ipmitool lan print can only print out
`Set in Progress` and other parameters are missing on our servers. After
bisecting, commit:

  351dad24a26f lan: Add processing of get/set specific CCs

is identified to be the source of the problem, where the function
get_lan_param_select is expected to consider severial error codes it
receives as empty response. It then constructs an empty response in `p`
and assigns `p` to `rc` and the control flow falls through to the
default case, which prints the error code in verbose mode and should
return `rc` instead of `NULL`.

Signed-off-by: Miao Wang <shankerwangmiao@gmail.com>